[fx] add shard utils.
